Initial Commit! Welcome to my Password Bank project...
The Main.java file and PasswordBank.java file were created along with a
bit of documentation at the start. The only thing in the main method
right now is a quick copy and paste from here:
https://www.javatpoint.com/java-jtextfield
Which this website will be crucial in teaching me more about java,
namely, the JTextField library.